requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
boto3==1.23.8
botocore==1.26.8; python_version >= '3.6'
jmespath==1.0.0; python_version >= '3.7'
numpy==1.22.4; python_version >= '3.8'
pandas==1.4.2
pyarrow==8.0.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.6
pytz==2022.1
pyyaml==6.0
s3transfer==0.5.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
